feat: 예약 시간 관리 API를 관리자 경로로 분리
