+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -1,23 +1,13 @@
 package roomescape.controller;
 
-import jakarta.annotation.Nonnull;
-import jakarta.validation.Valid;
-import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.dto.ReservationTimeAvailabilityResponseDto;
-import roomescape.dto.ReservationTimeRequestDto;
-import roomescape.dto.ReservationTimeResponseDto;
 import roomescape.service.ReservationTimeService;
 
 @RestController
@@ -29,31 +19,6 @@ public class ReservationTimeController {
         this.reservationTimeService = reservationTimeService;
     }
 
-    @PostMapping
-    public ResponseEntity<ReservationTimeResponseDto> create(@Valid @RequestBody ReservationTimeRequestDto request) {
-        ReservationTimeResponseDto response = reservationTimeService.create(request);
-
-        URI location = buildLocationUri(response);
-
-        return ResponseEntity.created(location)
-                .body(response);
-    }
-
-    @Nonnull
-    private static URI buildLocationUri(ReservationTimeResponseDto responseDto) {
-        return ServletUriComponentsBuilder
-                .fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(responseDto.id())
-                .toUri();
-    }
-
-    @GetMapping
-    public ResponseEntity<List<ReservationTimeResponseDto>> readAll() {
-        List<ReservationTimeResponseDto> responseDtos = reservationTimeService.readAll();
-        return ResponseEntity.ok(responseDtos);
-    }
-
     @GetMapping("/available")
     public ResponseEntity<List<ReservationTimeAvailabilityResponseDto>> readAvailabilityByDateAndTheme(
             @RequestParam("date") LocalDate date, @RequestParam("themeId") Long themeId) {
@@ -62,14 +27,5 @@ public class ReservationTimeController {
                 date, themeId);
 
         return ResponseEntity.ok(responseDtos);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-
-        reservationTimeService.delete(id);
-        return ResponseEntity
-                .noContent()
-                .build();
     }
 }
