+++ src/test/java/roomescape/acceptance/ReservationTimeAcceptanceTest.java
@@ -25,13 +25,13 @@ public class ReservationTimeAcceptanceTest {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(time)
-                .when().post("/times")
+                .when().post("/admin/times")
                 .then().log().all()
                 .statusCode(201);
 
         // 2. 전체 시간 조회
         RestAssured.given().log().all()
-                .when().get("/times")
+                .when().get("/admin/times")
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(1));
@@ -89,13 +89,13 @@ public class ReservationTimeAcceptanceTest {
 
         // 8. 시간 삭제
         RestAssured.given().log().all()
-                .when().delete("/times/1")
+                .when().delete("/admin/times/1")
                 .then().log().all()
                 .statusCode(204);
 
         // 9. 전체 시간 조회로 시간 삭제 확인
         RestAssured.given().log().all()
-                .when().get("/times")
+                .when().get("/admin/times")
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(0));
