+++ src/test/java/roomescape/acceptance/ThemeAcceptanceTest.java
@@ -44,7 +44,7 @@ public class ThemeAcceptanceTest {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(time)
-                .when().post("/times")
+                .when().post("/admin/times")
                 .then().log().all()
                 .statusCode(201);
